Move the bulk of the walk96 main back into the direwolf package
Issues using cgo structs across packages, plus just easier to have more
of the code together
